Try to fix 32bit php issue for IP address

+++ api/server.php
@@ -35,7 +35,7 @@ try
             {
                 $userid = isset($_POST['userid']) ? (int)$_POST['userid'] : 0;
                 $token = isset($_POST['token']) ? $_POST['token'] : "";
-                $address = isset($_POST['address']) ? (int)utf8_encode($_POST['address']) : null;
+                $address = isset($_POST['address']) ? utf8_encode($_POST['address']) : null;
                 $port = isset($_POST['port']) ? (int)utf8_encode($_POST['port']) : null;
                 $private_port = isset($_POST['private_port']) ? (int)utf8_encode($_POST['private_port']) : null;
                 $server_name = isset($_POST['name']) ? utf8_encode($_POST['name']) : "";
@@ -73,7 +73,7 @@ try
             {
                 $userid = isset($_POST['userid']) ? (int)utf8_encode($_POST['userid']) : 0;
                 $token = isset($_POST['token']) ? utf8_encode($_POST['token']) : null;
-                $address = isset($_POST['address']) ? (int)utf8_encode($_POST['address']) : 0;
+                $address = isset($_POST['address']) ? utf8_encode($_POST['address']) : 0;
                 $port = isset($_POST['port']) ? (int)utf8_encode($_POST['port']) : 0;
 
                 ClientSession::get($token, $userid)->stopServer($address, $port);
@@ -112,7 +112,7 @@ try
                 $userid = isset($_POST['userid']) ? (int)utf8_encode($_POST['userid']) : 0;
                 $token = isset($_POST['token']) ? utf8_encode($_POST['token']) : null;
                 $server_id = isset($_POST['server-id']) ? (int)utf8_encode($_POST['server-id']) : 0;
-                $address = isset($_POST['address']) ? (int)utf8_encode($_POST['address']) : 0;
+                $address = isset($_POST['address']) ? utf8_encode($_POST['address']) : 0;
                 $port = isset($_POST['port']) ? (int)utf8_encode($_POST['port']) : 0;
                 $aes_key = isset($_POST['aes-key']) ? utf8_encode($_POST['aes-key']) : null;
                 $aes_iv = isset($_POST['aes-iv']) ? utf8_encode($_POST['aes-iv']) : null;
@@ -136,7 +136,7 @@ try
             {
                 $userid = isset($_POST['userid']) ? (int)utf8_encode($_POST['userid']) : 0;
                 $token = isset($_POST['token']) ? utf8_encode($_POST['token']) : null;
-                $address = isset($_POST['address']) ? (int)utf8_encode($_POST['address']) : 0;
+                $address = isset($_POST['address']) ? utf8_encode($_POST['address']) : 0;
                 $port = isset($_POST['port']) ? (int)utf8_encode($_POST['port']) : 0;
                 $current_players = isset($_POST['current_players']) ? (int)$_POST['current_players'] : 0;
